Move stuff to demo app
