+++ src/accel/CMakeLists.txt
@@ -16,7 +16,6 @@ list(APPEND SOURCES
   ActionInitialization.cc
   EventAction.cc
   Logger.cc
-  PrimaryGeneratorAction.cc
   RunAction.cc
   TrackingAction.cc
 )
